+++ src/deluge/model/mod_controllable/GranularProcessor.h
@@ -1,0 +1,68 @@
+/*
+ * Copyright © 2024 Mark Adams and Alter-Alter
+ *
+ * This file is part of The Synthstrom Audible Deluge Firmware.
+ *
+ * The Synthstrom Audible Deluge Firmware is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#ifndef DELUGE_GRANULARPROCESSOR_H
+#define DELUGE_GRANULARPROCESSOR_H
+#include "GranularProcessor.h"
+#include "definitions_cxx.hpp"
+#include "dsp/stereo_sample.h"
+#include "hid/button.h"
+#include "modulation/lfo.h"
+#include "modulation/params/param.h"
+#include "modulation/params/param_descriptor.h"
+#include "modulation/params/param_set.h"
+
+struct Grain {
+	int32_t length;     // in samples 0=OFF
+	int32_t startPoint; // starttimepos in samples
+	int32_t counter;    // relative pos in samples
+	uint16_t pitch;     // 1024=1.0
+	int32_t volScale;
+	int32_t volScaleMax;
+	bool rev;        // 0=normal, 1 =reverse
+	int32_t panVolL; // 0 - 1073741823
+	int32_t panVolR; // 0 - 1073741823
+};
+
+class GranularProcessor {
+public:
+	GranularProcessor();
+	GranularProcessor(const GranularProcessor& other); // copy constructor
+	~GranularProcessor();
+	void setWrapsToShutdown();
+	// Grain
+	int32_t wrapsToShutdown;
+	StereoSample* modFXGrainBuffer{0};
+	uint32_t modFXGrainBufferWriteIndex;
+	int32_t grainSize;
+	int32_t grainRate;
+	int32_t grainShift;
+	Grain grains[8];
+	int32_t grainFeedbackVol;
+	int32_t grainVol;
+	int32_t grainDryVol;
+	int8_t grainPitchType;
+	bool grainLastTickCountIsZero;
+	bool grainInitialized;
+	void processGrainFX(StereoSample* buffer, int32_t modFXRate, int32_t modFXDepth, int32_t* postFXVolume,
+	                    UnpatchedParamSet* unpatchedParams, const StereoSample* bufferEnd);
+	void setupGrainFX(int32_t modFXRate, int32_t modFXDepth, int32_t* postFXVolume, UnpatchedParamSet* unpatchedParams);
+	void processOneGrainSample(StereoSample* currentSample);
+	void clearGrainFXBuffer();
+};
+
+#endif // DELUGE_GRANULARPROCESSOR_H
